label long to float32

+++ siamese_nn/data_prep.py
@@ -7,8 +7,8 @@ import numpy as np
 
 
 def dataset_and_split(x, y):
-    x = torch.tensor(np.array(x)) # list --> numpy --> torch is faster than list --> torch
-    y = torch.tensor(np.array(y))
+    x = torch.tensor(np.array(x))  # list --> numpy --> torch is faster than list --> torch
+    y = torch.tensor(np.array(y), dtype=torch.float32)  # label should be float32
     y = y.unsqueeze(dim=1)  # (5000) --> (5000,1)
     dataset = TensorDataset(x, y)
     train, validation, test = random_split(dataset, [3500, 1000, 500])
